added remove func to governance page
