+++ src/Components/Presentational/Table/TableWrapper.tsx
@@ -60,7 +60,8 @@ const TableWrapper = ({
   selected = [],
   setToggleAllCheckbox = null,
   setToggleCheckbox = null,
-  remove = false
+  remove = false,
+  setIsModalOpen = () => {}
 }: TableWrapperProps) => {
   const classes = useStyles();
 
@@ -152,6 +153,7 @@ const TableWrapper = ({
             orderBy={orderBy}
             rowCount={filteredData.length}
             selected={selected}
+            remove={remove}
           />
           {isLoading ? (
             <TableWrapperSkeleton />
@@ -168,6 +170,7 @@ const TableWrapper = ({
               page={page}
               rowsPerPage={rowsPerPage}
               remove={remove}
+              setIsModalOpen={setIsModalOpen}
             />
           ) : (
             <TableWrapperNotFoundFilter
@@ -213,18 +216,27 @@ interface TableWrapperProps {
   isLoading?: boolean;
   searchInput?: string;
   setFirstColLink?: Function;
-  // NOTE: filter requires setTitleText to work
+  /**
+   * Filter requires setTitleText to work
+   */
   setTitleText?: Function | null;
   filter?: Array<string> | null;
-  // NOTE: footerButtonLink and setFooterButtonClick requires footerButtonText to work
+  /**
+   * FooterButtonLink and setFooterButtonClick requires footerButtonText to work
+   */
   footerButtonText?: string;
   footerButtonLink?: string;
   setFooterButtonClick?: Function;
-  // NOTE: these 3 props need to all be there for checkboxes to work;
-  //       omitting all of them will remove all checkboxes
+  /**
+   * These 3 props need to all be there for checkboxes to work;
+   * omitting all of them will remove all checkboxes
+   */
   selected?: Array<any>;
   setToggleAllCheckbox?: Function | null;
   setToggleCheckbox?: Function | null;
-  // Remove
+  /**
+   * These 2 props need to all be there for Remove to work;
+   */
   remove?: boolean;
+  setIsModalOpen?: Function;
 }
